Sequence of images and storege improved
Now the sequence of images is set in a string variable. This change is with the aim to compute the total of images to project to know the index of images to ignore and acquire, and the index of the image where the camera trigger mode is turned off, dynamically. That is, knowing these index regardless of the number of images in the pattern sequence.